feat: Add KafkaConfig class for producing and consuming messages
The code changes include adding a new file `config.js` which defines a `KafkaConfig` class. This class is responsible for setting up the Kafka client, including the producer and consumer. It provides methods for producing messages to a specified topic and consuming messages from a specified topic. The `sendMessageToKafka` function in the `controller.js` file uses this class to produce messages to the "my-topic" topic.

The commit message suggests that a new feature has been added to the codebase, specifically the ability to configure and interact with Kafka for message production and consumption.